+++ modules/binary/impl/src/main/java/org/apache/ignite/internal/binary/builder/BinaryObjectArrayLazyValue.java
