+++ modules/binary/impl/src/main/resources/META-INF/services/org.apache.ignite.internal.binary.builder.BinaryObjectBuildersFactory
@@ -1,0 +1,1 @@
+org.apache.ignite.internal.binary.builder.BinaryObjectBuildersFactoryImpl
